Restore sensor names in TempData and HumidityData classes.
The sensor name in temperature and humidity channels was showing
as "undefined" in the web UI. This was due to the updated odin-control
ParameterTree implementation, which treats a 'name' parameter as part
of metadata and so does not return it without the metadata qualifier on
a request. Changed the parameter to 'sensor_name' in classes and also
in JS library, in lieu of a full metadata implementation.

Closes #47